+++ api/scripts/prod/create-profiles-collection-campaigns.js
